EVOLUTION : Search and PDF printing done

+++ src/app/controller/service/evolutions/echelle.service.ts
@@ -12,11 +12,16 @@ export class EchelleService {
   private _url = 'http://localhost:8099/evolution/echelle/';
   private _echelle: Echelle = new Echelle('', '', 0, '', null, '');
   private _echelles: Array<Echelle>;
+  private _searchInput: string;
+
+
 
   private SWAL_REACT = getReact('Echelle', true);
   private SUCCESS_SUCCESS_CREATE = this.SWAL_REACT.SUCCESS_CREATE;
   private SUCCESS_SUCCESS_EDIT = this.SWAL_REACT.SUCCESS_EDIT;
   private SUCCESS_SUCCESS_DELETE = this.SWAL_REACT.SUCCESS_DELETE;
+  private CONFIRMATION_DELETE_CONFIRMATION = this.SWAL_REACT.CONFIRMATION_DELETE_CONFIRMATION;
+  private SEARCH_NOT_FOUND = this.SWAL_REACT.SEARCH_NOT_FOUND;
   private ERROR_REF_ALREADY_EXISTS = this.SWAL_REACT.ERROR_REF_ALREADY_EXISTS;
   private ERROR_REF_DOES_NOT_EXIST = this.SWAL_REACT.ERROR_REF_DOES_NOT_EXIST;
   private ERROR_INVALID_REF = this.SWAL_REACT.ERROR_INVALID_REF;
@@ -50,6 +55,14 @@ export class EchelleService {
 
   set echelles(value: Array<Echelle>) {
     this._echelles = value;
+  }
+
+  get searchInput(): string {
+    return this._searchInput;
+  }
+
+  set searchInput(value: string) {
+    this._searchInput = value;
   }
 
   ajouterEchelle() {
@@ -98,19 +111,37 @@ export class EchelleService {
       });
   }
 
-  deleteEchelle(data){
-    this.http.delete(this._url + "delete/" + data).subscribe(
-      (res) => {
-        if (res == -1) {
-          Swal(this.ERROR_INVALID_REF);
-        } else if (res == -2) {
-          Swal(this.ERROR_REF_DOES_NOT_EXIST);
-        } else if (res == 1) {
+  deleteEchelle(data) {
+    Swal(this.CONFIRMATION_DELETE_CONFIRMATION)
+      .then((result) => {
+      if (result.value) {
+        this.http.delete(this._url + "delete/" + data).subscribe(
+          (res) => {
+            if (res == -1) {
+              Swal(this.ERROR_INVALID_REF);
+            } else if (res == -2) {
+              Swal(this.ERROR_REF_DOES_NOT_EXIST);
+            } else if (res == 1) {
+              this.getEchellesFromDatabase();
+              Swal(this.SUCCESS_SUCCESS_DELETE);
+            } else {
+              Swal(this.ERROR_UNKNOWN_ERROR);
+            }
+          });
+      }
+    });
+  }
+
+  search() {
+    if (this.searchInput !== "") {
+      this.echelles = this.echelles.filter(echelon => echelon.reference.includes(this.searchInput));
+      if (this.echelles === undefined || this.echelles.length == 0) {
+        Swal(this.SEARCH_NOT_FOUND).then(() => {
           this.getEchellesFromDatabase();
-          Swal(this.SUCCESS_SUCCESS_DELETE);
-        } else {
-          Swal(this.ERROR_UNKNOWN_ERROR);
-        }
-      });
+        });
+      }
+    }
   }
 }
+
+
