+++ src/app/controller/service/evolutions/echelon.service.ts
@@ -12,11 +12,14 @@ export class EchelonService {
   private _echelon = new Echelon('', 0, '');
   private _newEchelon = new Echelon('', 0, '');
   private _echelons = new Array<Echelon>();
+  private _searchInput: string;
 
   private SWAL_REACT = getReact('Echelon', false);
   private SUCCESS_SUCCESS_CREATE = this.SWAL_REACT.SUCCESS_CREATE;
   private SUCCESS_SUCCESS_EDIT = this.SWAL_REACT.SUCCESS_EDIT;
   private SUCCESS_SUCCESS_DELETE = this.SWAL_REACT.SUCCESS_DELETE;
+  private CONFIRMATION_DELETE_CONFIRMATION = this.SWAL_REACT.CONFIRMATION_DELETE_CONFIRMATION;
+  private SEARCH_NOT_FOUND = this.SWAL_REACT.SEARCH_NOT_FOUND;
   private ERROR_REF_ALREADY_EXISTS = this.SWAL_REACT.ERROR_REF_ALREADY_EXISTS;
   private ERROR_REF_DOES_NOT_EXIST = this.SWAL_REACT.ERROR_REF_DOES_NOT_EXIST;
   private ERROR_INVALID_REF = this.SWAL_REACT.ERROR_INVALID_REF;
@@ -60,6 +63,14 @@ export class EchelonService {
 
   set echelons(value: Echelon[]) {
     this._echelons = value;
+  }
+
+  get searchInput(): string {
+    return this._searchInput;
+  }
+
+  set searchInput(value: string) {
+    this._searchInput = value;
   }
 
   public ajouterEchelon() {
@@ -107,18 +118,34 @@ export class EchelonService {
   }
 
   supprimerEchelon(data) {
-    this.http.delete(this._url + "delete/" + data).subscribe(
-      (res) => {
-        if (res == -1) {
-          Swal(this.ERROR_INVALID_REF);
-        } else if (res == -2) {
-          Swal(this.ERROR_REF_DOES_NOT_EXIST);
-        } else if (res == 1) {
-          Swal(this.SUCCESS_SUCCESS_DELETE);
-          this.getEchelonsFromDatabase();
-        } else {
-          Swal(this.ERROR_UNKNOWN_ERROR);
+    Swal(this.CONFIRMATION_DELETE_CONFIRMATION)
+      .then((result) => {
+        if (result.value) {
+          this.http.delete(this._url + "delete/" + data).subscribe(
+            (res) => {
+              if (res == -1) {
+                Swal(this.ERROR_INVALID_REF);
+              } else if (res == -2) {
+                Swal(this.ERROR_REF_DOES_NOT_EXIST);
+              } else if (res == 1) {
+                this.getEchelonsFromDatabase();
+                Swal(this.SUCCESS_SUCCESS_DELETE);
+              } else {
+                Swal(this.ERROR_UNKNOWN_ERROR);
+              }
+            });
         }
       });
+  }
+
+  search() {
+    if (this.searchInput !== "") {
+      this.echelons = this.echelons.filter(echelon => echelon.reference.includes(this.searchInput));
+      if (this.echelons === undefined || this.echelons.length == 0) {
+        Swal(this.SEARCH_NOT_FOUND).then(() => {
+          this.getEchelonsFromDatabase();
+        });
+      }
+    }
   }
 }
