+++ src/app/controller/service/evolutions/evolution-personnel.service.ts
@@ -14,12 +14,15 @@ export class EvolutionPersonnelService {
   private _evolutionPersonnel = new EvolutionPersonnel('', null, '', '', null, null, '');
   private _newEvolutionPersonnel = new EvolutionPersonnel('', null, '', '', null, null, '');
   private _evolutionsPersonnel = new Array<EvolutionPersonnel>();
+  private _searchInput: string;
 
 
   private SWAL_REACT = getReact('Evolution personnel', true);
   private SUCCESS_SUCCESS_CREATE = this.SWAL_REACT.SUCCESS_CREATE;
   private SUCCESS_SUCCESS_EDIT = this.SWAL_REACT.SUCCESS_EDIT;
   private SUCCESS_SUCCESS_DELETE = this.SWAL_REACT.SUCCESS_DELETE;
+  private CONFIRMATION_DELETE_CONFIRMATION = this.SWAL_REACT.CONFIRMATION_DELETE_CONFIRMATION;
+  private SEARCH_NOT_FOUND = this.SWAL_REACT.SEARCH_NOT_FOUND;
   private ERROR_REF_ALREADY_EXISTS = this.SWAL_REACT.ERROR_REF_ALREADY_EXISTS;
   private ERROR_REF_DOES_NOT_EXIST = this.SWAL_REACT.ERROR_REF_DOES_NOT_EXIST;
   private ERROR_INVALID_REF = this.SWAL_REACT.ERROR_INVALID_REF;
@@ -72,6 +75,13 @@ export class EvolutionPersonnelService {
     );
   }
 
+  get searchInput(): string {
+    return this._searchInput;
+  }
+
+  set searchInput(value: string) {
+    this._searchInput = value;
+  }
 
   ajouterEvolutionPersonnel() {
     console.log(this._evolutionPersonnel);
@@ -108,18 +118,34 @@ export class EvolutionPersonnelService {
   }
 
   deleteEvolutionPersonnel(data) {
-    this.http.delete(this._url + "delete/" + data).subscribe(
-      (res) => {
-        if (res == -1) {
-          Swal(this.ERROR_INVALID_REF);
-        } else if (res == -2) {
-          Swal(this.ERROR_REF_DOES_NOT_EXIST);
-        } else if (res == 1) {
-          this.getEvolutionsPersonnelFromDatabase();
-          Swal(this.SUCCESS_SUCCESS_DELETE);
-        } else {
-          Swal(this.ERROR_UNKNOWN_ERROR);
+    Swal(this.CONFIRMATION_DELETE_CONFIRMATION)
+      .then((result) => {
+        if (result.value) {
+          this.http.delete(this._url + "delete/" + data).subscribe(
+            (res) => {
+              if (res == -1) {
+                Swal(this.ERROR_INVALID_REF);
+              } else if (res == -2) {
+                Swal(this.ERROR_REF_DOES_NOT_EXIST);
+              } else if (res == 1) {
+                this.getEvolutionsPersonnelFromDatabase();
+                Swal(this.SUCCESS_SUCCESS_DELETE);
+              } else {
+                Swal(this.ERROR_UNKNOWN_ERROR);
+              }
+            });
         }
       });
+  }
+
+  search() {
+    if (this.searchInput !== "") {
+      this.evolutionsPersonnel = this.evolutionsPersonnel.filter(echelon => echelon.reference.includes(this.searchInput));
+      if (this.evolutionsPersonnel === undefined || this.evolutionsPersonnel.length == 0) {
+        Swal(this.SEARCH_NOT_FOUND).then(() => {
+          this.getEvolutionsPersonnelFromDatabase();
+        });
+      }
+    }
   }
 }
