+++ src/app/controller/service/evolutions/loi-evolution.service.ts
@@ -14,11 +14,14 @@ export class LoiEvolutionService {
   private _url = 'http://localhost:8099/evolution/loi/';
   private _loiEvolution = new LoiEvolution('', '', '', '');
   private _loisEvolution = new Array<LoiEvolution>();
+  private _searchInput: string;
 
   private SWAL_REACT = getReact('Loi d\'evolution personnel', true);
   private SUCCESS_SUCCESS_CREATE = this.SWAL_REACT.SUCCESS_CREATE;
   private SUCCESS_SUCCESS_EDIT = this.SWAL_REACT.SUCCESS_EDIT;
   private SUCCESS_SUCCESS_DELETE = this.SWAL_REACT.SUCCESS_DELETE;
+  private CONFIRMATION_DELETE_CONFIRMATION = this.SWAL_REACT.CONFIRMATION_DELETE_CONFIRMATION;
+  private SEARCH_NOT_FOUND = this.SWAL_REACT.SEARCH_NOT_FOUND;
   private ERROR_REF_ALREADY_EXISTS = this.SWAL_REACT.ERROR_REF_ALREADY_EXISTS;
   private ERROR_REF_DOES_NOT_EXIST = this.SWAL_REACT.ERROR_REF_DOES_NOT_EXIST;
   private ERROR_INVALID_REF = this.SWAL_REACT.ERROR_INVALID_REF;
@@ -52,6 +55,14 @@ export class LoiEvolutionService {
 
   set loisEvolution(value: LoiEvolution[]) {
     this._loisEvolution = value;
+  }
+
+  get searchInput(): string {
+    return this._searchInput;
+  }
+
+  set searchInput(value: string) {
+    this._searchInput = value;
   }
 
   public getLoisEvolutionsFromDatabase() {
@@ -94,20 +105,34 @@ export class LoiEvolutionService {
   }
 
   deleteLoiEvolution(data) {
-    this.http.delete(this._url + "delete/" + data).subscribe(
-      (res) => {
-        if (res == -1) {
-          Swal(this.ERROR_INVALID_REF);
-        } else if (res == -2) {
-          Swal(this.ERROR_REF_DOES_NOT_EXIST);
-        } else if (res == 1) {
-          this.getLoisEvolutionsFromDatabase();
-          Swal(this.SUCCESS_SUCCESS_DELETE);
-        } else {
-          Swal(this.ERROR_UNKNOWN_ERROR);
+    Swal(this.CONFIRMATION_DELETE_CONFIRMATION)
+      .then((result) => {
+        if (result.value) {
+          this.http.delete(this._url + "delete/" + data).subscribe(
+            (res) => {
+              if (res == -1) {
+                Swal(this.ERROR_INVALID_REF);
+              } else if (res == -2) {
+                Swal(this.ERROR_REF_DOES_NOT_EXIST);
+              } else if (res == 1) {
+                this.getLoisEvolutionsFromDatabase();
+                Swal(this.SUCCESS_SUCCESS_DELETE);
+              } else {
+                Swal(this.ERROR_UNKNOWN_ERROR);
+              }
+            });
         }
       });
   }
 
-
+  search() {
+    if (this.searchInput !== "") {
+      this.loisEvolution = this.loisEvolution.filter(echelon => echelon.reference.includes(this.searchInput));
+      if (this.loisEvolution === undefined || this.loisEvolution.length == 0) {
+        Swal(this.SEARCH_NOT_FOUND).then(() => {
+          this.getLoisEvolutionsFromDatabase();
+        });
+      }
+    }
+  }
 }
